requirements.txt
tensorflow
gunicorn
pillow
flask
numpy
keras
